requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.21.6
openai==0.23.0
openpyxl==3.0.10
pandas==1.3.5
pandas-stubs==1.2.0.62
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
telegram.py==0.3.1
tornado==6.2
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
